deep/requirements.txt
numpy
torch==1.4.0
ipdb
gym
atari_py
gym[all]
pyyaml
dill
opencv-python